backend/app/scripts/data_ingestion/html_scraper.py
# Placeholder for html_scraper.py
import requests
import time
from urllib.robotparser import RobotFileParser
from bs4 import BeautifulSoup

def scrape_website(url: str, user_agent: str = "AutoModelerBot/1.0 (+http://automodeler.example.com/bot.html)"):
    """
    Scrapes a website, respecting its robots.txt file.
    """
    # 1. Check robots.txt
    rp = RobotFileParser()
    rp.set_url(f"{url.scheme}://{url.netloc}/robots.txt")
    rp.read()
    if not rp.can_fetch(user_agent, url.geturl()):
        print(f"Scraping disallowed by robots.txt for {url.geturl()}")
        return None

    # 2. Scrape with rate limiting
    print(f"Scraping allowed, fetching {url.geturl()}...")
    headers = {"User-Agent": user_agent}
    try:
        time.sleep(2) # Rate limit
        response = requests.get(url.geturl(), headers=headers)
        response.raise_for_status()
        
        # Basic parsing example
        soup = BeautifulSoup(response.content, 'html.parser')
        return soup.title.string

    except requests.exceptions.RequestException as e:
        print(f"Error fetching {url.geturl()}: {e}")
        return None